Make version number optional and reflect in help
This commit modifies the BasicHelpFunc to check whether Version is an
empty string, and if it is omits it from the help generated by
BasicHelpFunc. It also adds a test of this behaviour.

This is useful in situations where you are using nested CLIs to get
aws-cli style user interfaces, and conseuqently only want Version as a
top level switch - for example:

    aws ec2 --help

Should not list --version as an option, but

    aws --help

Should list --version as an option.