Add formatting functions
This commit implements *f functions such as Errorf and Askf on all the
Ui objects as well as the interface itself. This allows users to
simplify their implementation code from:

    c.ui.Warn(fmt.Sprintf("Invalid token %q", token))

to

    c.ui.Warnf("Invalid token %q", token)

This is based off of the work @tarrant did in #13, except I elected to
add it directly to the interface instead of making a new interface type.

I don't think this breaks compatability unless someone is implementing
their own Ui, since it updates the interface.

@mitchellh do you think this will break over RPC?