mockui: fix data race and nil panic
This patch fixes two issues with MockUi which result either in a nil
panic or a data race. The nil panic happens when the no output was
written before the String() method is called. Since the buffers are
initialized on the first write there is no chance to initialize the
buffers. With the given design there is no way to fix this without
breaking existing code. Users are encouraged to use the NewMockUi method
which returns a fully initialized MockUi instead of just creating the
MockUi instance manually.

Additionally, the output buffers are not synchronized which triggers a
data race between readers and writers of the buffers. The buffers have
been replaced with synchronized buffers to address this problem.