Root help should only print root subcommands
When the root help is activated (subcommand is "" and help flag is
specified), we should only show the root subcommands.

This is likely from when we added subcommand support years ago and never
actually edited the root help to use the `helpCommands` helper to only
return the immediate subcommands for a command.