Only look for subcommands up to quoted with spaces
Fixes #68

This changes the subcommand search behavior to only look up to the point
where a space is found in a single arg. This doesn't break prior
behavior because prior behavior would've resulted in a panic.

Examples:

1. `./cli foo "bar baz"` would match subcommand "foo" and treat "bar baz"
   as a single argument to subcommand "foo".

2. `./cli "foo bar"` would have no subcommand.

The reason I took this approach is because we otherwise treat quoted
arguments with a space as a single argument (as do shells). This keeps
that behavior in line.