exclude flags from subcommand parsing
When determining the string to use when querying the radix tree for the
longest subcommand, exclude any argument beginning with a "-", since
this is a flag and not a subcommand.

This fixes a bug in which "fooasdf -o=foo" is matched to the subcommand
"foo" due to logic further down on line 704.